+++ RebalancedIndustriesRevisited/Extensions/FishFarmAIExtensions.cs
@@ -2,10 +2,10 @@
 
 namespace RebalancedIndustriesRevisited.Extensions;
 
-public static class IndustryBuildingAIExtension {
-    public static WorkPlace GetWorkPlace(this IndustryBuildingAI ai) => new(ai.m_workPlaceCount0, ai.m_workPlaceCount1, ai.m_workPlaceCount2, ai.m_workPlaceCount3);
+public static class FishFarmAIExtensions {
+    public static WorkPlace GetWorkPlace(this FishFarmAI ai) => new(ai.m_workPlaceCount0, ai.m_workPlaceCount1, ai.m_workPlaceCount2, ai.m_workPlaceCount3);
 
-    public static void SetWorkPlace(this IndustryBuildingAI ai, WorkPlace workPlace) {
+    public static void SetWorkPlace(this FishFarmAI ai, WorkPlace workPlace) {
         ai.m_workPlaceCount0 = workPlace.UneducatedWorkers;
         ai.m_workPlaceCount1 = workPlace.EducatedWorkers;
         ai.m_workPlaceCount2 = workPlace.WellEducatedWorkers;